% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm0010_equations.R
\name{calculate_leakage_emissions_acm0010}
\alias{calculate_leakage_emissions_acm0010}
\title{Leakage emissions for ACM0010 projects}
\usage{
calculate_leakage_emissions_acm0010(
  electricity_consumed_mwh,
  grid_emission_factor,
  additional_leakage = 0
)
}
\arguments{
\item{electricity_consumed_mwh}{Electricity consumed from the grid in MWh.}

\item{grid_emission_factor}{Emission factor of the connected grid in tCO2e per
MWh.}

\item{additional_leakage}{Additional leakage emissions in tonnes CO2e to add
(e.g. transport of digestate).}
}
\value{
Numeric vector of leakage emissions in tonnes CO2e.
}
\description{
Calculates emissions associated with auxiliary electricity consumption or fuel
use that are attributable to the project.
}
\examples{
calculate_leakage_emissions_acm0010(50, 0.8, 5)
}
