% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm0010_equations.R
\name{convert_methane_volume_to_mass}
\alias{convert_methane_volume_to_mass}
\title{Convert methane volume to mass}
\usage{
convert_methane_volume_to_mass(volume_m3, methane_density = 0.0007168)
}
\arguments{
\item{volume_m3}{Numeric vector of methane volumes in cubic metres.}

\item{methane_density}{Density of methane in tonnes per cubic metre. Defaults
to \code{0.0007168}, equivalent to 0.7168 kg/m3 at standard temperature and
pressure.}
}
\value{
Numeric vector of methane mass in tonnes.
}
\description{
Translates methane volume expressed at standard conditions into tonnes of
methane using a default density of 0.0007168 tonnes per cubic metre.
}
\examples{
convert_methane_volume_to_mass(100)
convert_methane_volume_to_mass(c(50, 75), methane_density = 0.0007)
}
