% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm0010_ontology.R
\name{read_acm0010_technology_shapes}
\alias{read_acm0010_technology_shapes}
\title{Load ACM0010 technology applicability shape}
\usage{
read_acm0010_technology_shapes()
}
\value{
A \code{sh_shape_graph} as returned by \code{\link[shaclR:read_shacl]{shaclR::read_shacl()}}.
}
\description{
Load ACM0010 technology applicability shape
}
